Update test for new API
