+++ notebooks/sample_data/sample_data.vol.info
@@ -1,0 +1,12 @@
+! PyHST_SLAVE VOLUME INFO FILE
+NUM_X =  200
+NUM_Y =  200
+NUM_Z =  200
+voxelSize =  1.000000e-02
+BYTEORDER = LOWBYTEFIRST
+ValMin =  -2.04782312e+03 
+ValMax =  2.03446960e+03 
+s1 =  -1.73709910e+03 
+s2 =  1.64998256e+03 
+S1 =  -1.45101174e+03 
+S2 =  1.39763538e+03
